backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
strawberry-graphql==0.248.0
aioinject==0.38.0

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Settings & Config
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# Logging
loguru==0.7.2

# Data Processing
pandas==2.2.3

# LangChain & AI
langchain==0.3.10
langchain-core==0.3.21
langchain-google-vertexai==2.0.11
langchain-chroma==0.1.4

# Google Cloud
google-cloud-aiplatform==1.70.0
google-auth==2.36.0

# Redis (opcional para caché)
redis==5.2.0

# Testing (opcional)
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
